Replacing test in a more logical file
